fix upload bugs
