+++ packages/arcgis-rest-portal/src/items/add.ts
@@ -10,8 +10,7 @@ import {
   IUpdateItemResponse,
   IItemResourceResponse,
   determineOwner,
-  IManageItemRelationshipOptions,
-  IItemPartOptions
+  IManageItemRelationshipOptions
 } from "./helpers";
 import { updateItem, IUpdateItemOptions } from "./update";
 
@@ -135,41 +134,5 @@ export function addItemResource(
     };
 
     return request(url, requestOptions);
-  });
-}
-
-/**
- * ```js
- * import { addItemPart } from "@esri/arcgis-rest-portal";
- * //
- * addItemPart({
- *   id: "30e5fe3149c34df1ba922e6f5bbf808f",
- *   part: data,
- *   partNum: 1,
- *   authentication
- * })
- *   .then(response)
- * ```
- * Inquire about status when publishing an item, adding an item in async mode, or adding with a multipart upload. See the [REST Documentation](https://developers.arcgis.com/rest/users-groups-and-items/add-item-part.htm) for more information.
- *
- * @param id - The Id of the item to get status for.
- * @param requestOptions - Options for the request
- * @returns A Promise to get the item status.
- */
-export function addItemPart(
-  requestOptions?: IItemPartOptions
-): Promise<IUpdateItemResponse> {
-  return determineOwner(requestOptions).then(owner => {
-    const url = `${getPortalUrl(requestOptions)}/content/users/${owner}/items/${
-      requestOptions.id
-    }/addPart`;
-
-    const options = appendCustomParams<IItemPartOptions>(
-      requestOptions,
-      ["file", "partNum"],
-      { params: { ...requestOptions.params } }
-    );
-
-    return request(url, options);
   });
 }
